requirements.txt
celery==5.2.7
w3lib==2.0.1
parsel==1.6.0
pandas==1.4.4
matplotlib==3.5.3
networkx==2.8.6
httpx==0.23.0
scikit-learn==1.1.3
fastapi==0.91.0
urllib3==1.26.14
SQLAlchemy==2.0.3
pymysql==1.0.2
PyYAML==6.0
